Refactor file structure for posts
