+++ imports/api/posts/both/index.js
@@ -1,0 +1,1 @@
+import "./methods/posts";
