+++ imports/api/posts/client/index.js
@@ -1,0 +1,1 @@
+import "../both";
