+++ imports/api/posts/server/index.js
@@ -1,0 +1,3 @@
+import "./seed";
+import "./publications";
+import "../both";
