+++ imports/api/posts/server/publications.js
@@ -1,0 +1,6 @@
+import {Meteor} from "meteor/meteor";
+import {PostsCollection} from "../../../db/posts/collection";
+
+Meteor.publish('posts', function(){
+    return PostsCollection.find({});
+})
