+++ imports/db/posts/collection.js
@@ -1,0 +1,3 @@
+import {Mongo} from "meteor/mongo";
+
+export const PostsCollection = new Mongo.Collection('posts');
