+++ imports/ui/pages/posts/SinglePostPage.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { withTracker } from 'meteor/react-meteor-data';
-import { PostsCollection } from '../../../api/collections/postCollections';
 import { useHistory } from 'react-router-dom';
+import {PostsCollection} from "../../../db/posts/collection";
 
 
 let  email = "jordytshibss@test.com";
